requirements.txt
aniso8601==9.0.1
async-timeout==4.0.3
black==23.11.0
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-And-Redis==1.0.0
Flask-HTTPAuth==4.8.0
Flask-RESTful==0.3.10
idna==3.6
isort==5.13.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
Werkzeug==3.0.1
